[build] [refactor] Move wasm and dx11 source files to target locations (#5235)
* Split wasm source

* add dep

* Build examples depends on llvma and metal

* Define codegen target for metal and opengl

* Split dx11 source files

* Include llvm dir

* Codegen llvm include

* Depend on codegen
