+++ cmake/TaichiExamples.cmake
@@ -22,6 +22,11 @@ if (WIN32)
     set_target_properties(${EXAMPLES_NAME} PROPERTIES RUNTIME_OUTPUT_DIRECTORY_RELWITHDEBINFO ${EXAMPLES_OUTPUT_DIR})
 endif()
 target_link_libraries(${EXAMPLES_NAME} PRIVATE taichi_isolated_core)
+target_link_libraries(${EXAMPLES_NAME} PRIVATE
+	metal_program_impl
+	metal_runtime
+	metal_codegen
+	)
 
 # TODO 4832: be specific on the header dependencis here, e.g., ir
 target_include_directories(${EXAMPLES_NAME}
