+++ taichi/codegen/wasm/CMakeLists.txt
@@ -1,0 +1,15 @@
+# ./taichi/codegen/wasm/CMakeLists.txt
+
+add_library(wasm_codegen)
+target_sources(wasm_codegen
+  PRIVATE
+    codegen_wasm.cpp
+  )
+
+target_include_directories(wasm_codegen
+  PRIVATE
+    ${PROJECT_SOURCE_DIR}
+    ${PROJECT_SOURCE_DIR}/external/spdlog/include
+    ${PROJECT_SOURCE_DIR}/external/eigen
+    ${LLVM_INCLUDE_DIRS}
+  )
